requirements.txt
backcall==0.2.0
cffi==1.14.3
cryptography==3.1
decorator==4.4.2
Django==1.11.7
django-ranged-response==0.2.0
django-redis==4.11.0
django-redis-cache==2.1.1
django-simple-captcha==0.5.12
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycparser==2.20
pycryptodomex==3.9.4
Pygments==2.6.1
PyMySQL==0.10.0
pyOpenSSL==19.1.0
python-alipay-sdk==2.1.1
pytz==2020.1
redis==3.5.3
six==1.15.0
traitlets==4.3.3
wcwidth==0.2.5
